+++ firmware/m5/components/spi_bus/test/test_spi_bus.cpp
@@ -13,6 +13,8 @@
 
 #include <unity.h>
 
+#include "board.h"
+
 #include "spi_bus.h"
 
 namespace {